requirements.txt
black==23.1.0
click==8.1.3
colorama==0.4.6
flake8==6.0.0
markdown-it-py==2.1.0
mccabe==0.7.0
mdurl==0.1.2
mypy==0.991
mypy-extensions==0.4.3
mysql-connector-python==8.0.32
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
platformdirs==2.6.2
protobuf==3.20.3
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
rich==13.3.1
six==1.16.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
